Limit beaker-puppet to older Rubies
No version of beaker-puppet is compatible with Ruby >= 3.2 because of
taint functions. Because beaker-puppet specifies that it cannot run on
Ruby > 3.1 in its gemspec, when you run bundle install on newer
Rubies-even when excluding the Gem group to which beaker-puppet
belongs-Bundler errors.

This commit updates the Gemfile to only list beaker-puppet if Ruby is
less than 3.1.0.