(PA-5807) Ignore yumrepo files managed by Puppet file resource
Previously we attempted to manage every repo that was found in the yumrepos
directory. If one of those repos was a Puppet File resource, yumrepo
would manage the yumrepo the Puppet File resource was creating. This casued
some idepontecy issues.
This change will make it so yumrepo will not attempt to manage any file resource
it finds that is found in the yumrepo directory.